# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Backend API",
    "version": "1.0.0",
    "description": "Documentação da API do backend"
  },
  "servers": [
    {
      "url": "https://tematicoapi.zapchatbr.com/",
      "description": "API Temático"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/eventos": {
      "get": {
        "summary": "Listar eventos",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "data",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de eventos"
          }
        },
        "tags": [
          "Eventos"
        ]
      },
      "post": {
        "summary": "Criar evento",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string"
                  },
                  "data": {
                    "type": "string",
                    "format": "date"
                  },
                  "hora": {
                    "type": "string",
                    "pattern": "^\\d{2}:\\d{2}$"
                  },
                  "restauranteId": {
                    "type": "integer"
                  }
                },
                "required": [
                  "nome",
                  "data",
                  "hora",
                  "restauranteId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Evento criado"
          }
        },
        "tags": [
          "Eventos"
        ]
      }
    },
...
    "/eventos/{id}": {
      "get": {
        "summary": "Obter evento pelo ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento encontrado"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        },
        "tags": [
          "Eventos"
        ]
      },
      "put": {
        "summary": "Atualizar evento",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string"
                  },
                  "data": {
                    "type": "string",
                    "format": "date"
                  },
                  "hora": {
                    "type": "string"
                  },
                  "restauranteId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evento atualizado"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        },
        "tags": [
          "Eventos"
        ]
      },
      "delete": {
        "summary": "Deletar evento",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento deletado"
          },
          "400": {
            "description": "Evento possui reservas ativas"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        },
        "tags": [
          "Eventos"
        ]
      }
    },
...
    "/eventos-reservas": {
      "get": {
        "summary": "Listar marcações de eventos",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de marcações"
          }
        },
        "tags": [
          "EventosReservas"
        ]
      },
      "post": {
        "summary": "Criar marcação de evento",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "eventoId": {
                    "type": "integer"
                  },
                  "reservaId": {
                    "type": "integer"
                  },
                  "informacoes": {
                    "type": "string"
                  },
                  "quantidade": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string"
                  }
                },
                "required": [
                  "eventoId",
                  "reservaId",
                  "quantidade"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Marcação criada"
          }
        },
        "tags": [
          "EventosReservas"
        ]
      }
    },
    "/eventos-reservas/{eventoId}/{reservaId}": {
      "get": {
        "summary": "Obter marcação específica",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "eventoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "reservaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Marcação encontrada"
          },
          "404": {
            "description": "Marcação não encontrada"
          }
        },
        "tags": [
          "EventosReservas"
        ]
      },
      "put": {
        "summary": "Atualizar marcação",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "eventoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "reservaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "informacoes": {
                    "type": "string"
                  },
                  "quantidade": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Marcação atualizada"
          },
          "404": {
            "description": "Marcação não encontrada"
          }
        },
        "tags": [
          "EventosReservas"
        ]
      },
      "delete": {
        "summary": "Remover marcação",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "eventoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "reservaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Marcação removida"
          },
          "404": {
            "description": "Marcação não encontrada"
          }
        },
        "tags": [
          "EventosReservas"
        ]
      }
    },
    "/reservas": {
      "get": {
        "summary": "Listar reservas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "codigoUH",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hoje",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de reservas"
          }
        },
        "tags": [
          "Reservas"
        ]
      },
      "post": {
        "summary": "Criar reserva",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idreservacm": {
                    "type": "integer"
                  },
                  "numeroreservacm": {
                    "type": "string"
                  },
                  "coduh": {
                    "type": "string"
                  },
                  "nome_hospede": {
                    "type": "string"
                  },
                  "contato": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "data_checkin": {
                    "type": "string",
                    "format": "date"
                  },
                  "data_checkout": {
                    "type": "string",
                    "format": "date"
                  },
                  "qtd_hospedes": {
                    "type": "integer"
                  }
                },
                "required": [
                  "coduh",
                  "nome_hospede",
                  "contato",
                  "email",
                  "data_checkin",
                  "data_checkout",
                  "qtd_hospedes"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reserva criada"
          }
        },
        "tags": [
          "Reservas"
        ]
      }
    },
...
    "/reservas/{id}": {
      "put": {
        "summary": "Atualizar reserva",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idreservacm": {
                    "type": "integer"
                  },
                  "numeroreservacm": {
                    "type": "string"
                  },
                  "coduh": {
                    "type": "string"
                  },
                  "nome_hospede": {
                    "type": "string"
                  },
                  "contato": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "data_checkin": {
                    "type": "string",
                    "format": "date"
                  },
                  "data_checkout": {
                    "type": "string",
                    "format": "date"
                  },
                  "qtd_hospedes": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reserva atualizada"
          },
          "404": {
            "description": "Reserva não encontrada"
          }
        },
        "tags": [
          "Reservas"
        ]
      },
      "delete": {
        "summary": "Deletar reserva",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reserva deletada"
          },
          "404": {
            "description": "Reserva não encontrada"
          }
        },
        "tags": [
          "Reservas"
        ]
      }
    }
...
  }
}